DEV-2 small refactor by splitting Home Page into smaller components
