+++ src/components/ProfileData.vue
@@ -1,0 +1,22 @@
+<template>
+    <div class="panel panel-default">
+        <div class="panel-heading">
+            <h3 class="panel-title">Profile Data</h3>
+        </div>
+        <div class="panel-body text-center">
+            <img src="../assets/user_avatar.png" alt="avatar" class="img-rounded user-avatar">
+            <h2>Jordan Belfort<br/><small>Stratton Oakmont</small></h2>
+            <p>The Bronx, New York City</p>
+            <p class="text-muted">July 9, 1962</p>
+        </div>
+    </div>
+</template>
+
+<style lang="scss" scoped>
+    .user-avatar {
+      display: block;
+      margin: auto;
+      width: 100%;
+      max-width: 300px;
+    }
+</style>
